Issue 10 Move delete test case logic to Node.svelte
